requirements.txt
psycopg2
pyjwt
hashlib
csv
djangorestframework